requirements.txt
customtkinter
tkcalendar
pillow
aiohttp
requests
pandas
openpyxl
xlrd
python-dateutil
pdfplumber
pypdf
python-dotenv
